feat: improve timeout precision for read and write operations (#68)
- Change int timeout to float for all read operations (readMessage, readFrame, readLoop)
- Add optional write timeout to sendFrame and sendMessage
- Update Consumer::read/readOne to accept float timeout
- Update Producer::send/sendBatch to accept optional write timeout
- Update Producer::waitForConfirms to use float timeout with microtime precision
- Use consistent RuntimeException for all timeout errors
- Add tests for float timeout signatures and write timeout passthrough

All timeout parameters now accept float values (e.g. 0.5 for 500ms)
for sub-second precision. Write operations support optional timeout
to limit blocking time when socket is not ready.

+++ src/Client/Connection.php
@@ -190,7 +190,7 @@ class Connection
         return new Consumer($this->streamConnection, $stream, $subscriptionId, $offset, $name, $autoCommit, $initialCredit);
     }
 
-    public function readLoop(?int $maxFrames = null, ?int $timeout = null): void
+    public function readLoop(?int $maxFrames = null, ?float $timeout = null): void
     {
         $this->streamConnection->readLoop($maxFrames, $timeout);
     }
